Change to working URL

+++ download_pngsuite.sh
@@ -1,11 +1,11 @@
 #!/bin/sh
 
-URL="http://www.schaik.com/pngsuite2011/PngSuite-2017jul19.tgz"
+URL="https://sleepinginsomniac.github.io/png/png-suite.tbz"
 DEST="spec/fixtures"
 
 mkdir -p tmp
-curl -L $URL -o tmp/png-suite.tgz
-mkdir -p "$DEST/png-suite"
-tar -xzf tmp/suit.tgz -C "$DEST/png-suite"
-rm tmp/png-suite.tgz
-rm "$DEST/png-suite/PngSuite.png"
+if [ ! -f "tmp/png-suite.tbz" ]; then
+  curl -sL $URL -o "tmp/png-suite.tbz"
+fi
+tar -xjf "tmp/png-suite.tbz" -C $DEST
+[ -f "$DEST/png-suite/PngSuite.png" ] && rm "$DEST/png-suite/PngSuite.png"
